+++ thermion_dart/native/include/scene/SceneAsset.hpp
@@ -9,6 +9,7 @@
 
 #include "CustomGeometry.hpp"
 #include "Log.hpp"
+#include "c_api/APIBoundaryTypes.h"
 
 namespace thermion {
 
@@ -55,6 +56,10 @@ class SceneAsset {
         virtual Entity findEntityByName(const char* name) = 0;
 
         virtual const filament::Aabb getBoundingBox() const = 0;
+
+        virtual uint32_t getGeometryCapabilities() const {
+            return SCENE_ASSET_GEOMETRY_CAPABILITY_NONE;
+        }
 
         virtual size_t getBoneCount(size_t skinIndex) const {
             return 0;